Move MakeNewExtensions.cs to Factories folder in Shared project.
