+++ src/ExtensionManager.Shared/ExtensionManager.Shared.projitems
@@ -13,6 +13,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)Commands\ExportSolutionCommand.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Commands\ImportCommand.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Commands\SolutionPrompter.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Factories\MakeNewExtension.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Models\Extension.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Services\ExtensionService.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Importer\ImportWindow.xaml.cs">
